pyproject.toml
[tool.poetry]
name = "util"
version = "0.1.0"
description = "Python functions and classes for core utilities"
authors = ["Roger Chylla <rogerchylla1205@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
pymongo = "^4.6.1"
pyyaml = "^6.0.1"
numpy = "^1.25"

[tool.poetry.scripts]
test_connect = "util.scripts.test_connection:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
